Report on Height Based Entry System
*On certain roads of cities, there are a large number of overhead passes in highways. This makes it difficult for tall vehicles to pass through and it also results in traffic as the vehicle should now change the route. This sensor can be used in roads before overhead passes to check vehicles and to alert the 
driver to change routes if the vehicle exceeds the limit height.
*In amusement parks to check each person’s height before letting them in on certain rides as it can be dangerous to people with short height. This requires a large  number of safety personnel to be present at each ride and also results in delay in movement of people in queues.
Our project aims at providing solutions for these problems by using sensors before the entrance to check for height requirements and alert if the person/vehicle does not meet the required height. This reduces the man power needed to check people’s heights at rides in amusement parks and 
also saves time of a lot a people.