Single Port RAM Using VHDL Code
Our Project is Based on one of the type of RAM i.e., Single Port RAM . The Main purpose of Our Project is to study on working and components of Single Port RAM. Since the modern world is improving with improvement in technology, There is still backlogged in speed and accuracy in data retrieval in some field. Even though Single Port RAM is older compared to present world, The speed and accuracy is still best when compared to DRAM or any other normal RAM (Since DRAM acts as Sync ,Hence its speed is quite low Compared to SRAM).