Vehicle Movement Based Street Lights Using ATMEGA
Vehicle movement based street light is an automated system which automates the street light. The main aim of our project is to reduce the power consumption when there are no vehicle movements on the road. The street light will glow with high intensity when there are vehicles on the road otherwise the lights will remain OFF. With advancement of technology, things are becoming simpler and easier for everyone in the world today. Automation is the use of control systems and information technologies to reduce the need for human work in the production of goods and services in society.